minecraft_autoclicker/requirements.txt
pynput>=1.7.6
customtkinter>=5.2.0
pywin32>=306
Pillow>=10.0.1
pystray>=0.19.4
